\name{ allExps }
\alias{ allExps }
\title{ A SimpleList containing some dummy Expression Atlas datasets }
\description{
    This is a SimpleList object containing dummy data from some Expression Atlas
    experiments, to demonstrate a typical example of the reesults of using the
    \code{getAtlasData()} function. The dummy data objects only contain the
    first 10 rows of the real data, to save disk space. To get the real data,
    download it using \code{getAtlasData()}.
}
\usage{
    data( "allExps" )
}
\value{
    A SimpleList with one element per Expression Atlas dataset.
}
