\name{ atlasRes }
\alias{ atlasRes }
\title{ A DataFrame listing some Expression Atlas experiments }
\description{
    This is a DataFrame object listing some Expression Atlas experiments, to
    demonstrate a typical example of the reesults of using the
    \code{searchAtlasExperiments()} function.
}
\usage{
    data( "atlasRes" )
}
\value{
    A DataFrame with one row per Expression Atlas dataset.
}
